refactor: Move metrics to its own definition
